REFACTOR changed methods from using Integer to Long for idPizza

+++ src/main/java/it/itsacademy/pizzeriaexpress/service/PizzaService.java
@@ -5,9 +5,9 @@ import it.itsacademy.pizzeriaexpress.dto.PizzaDTO;
 public interface PizzaService {
     public PizzaDTO creaPizza(PizzaDTO nuovaPizza);
 
-    public PizzaDTO eliminaPizza(Integer idPizza);
+    public PizzaDTO eliminaPizza(Long idPizza);
 
-    public PizzaDTO cercaPizza(Integer idPizza);
+    public PizzaDTO cercaPizza(Long idPizza);
 
     public PizzaDTO modificaPizza(PizzaDTO nuovaPizza);
 }
